pyproject.toml
[tool.poetry]
name = "memory"
version = "0.2.2"
description = "papr memory server"
authors = ["Shawkat Kabbara <shawkat@papr.it>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "3.0.3"
flask-cors = "5.0.0"
python-jose = "3.3.0"
flask-login = "0.6.3"
python-dotenv = "1.0.1"
authlib = "1.3.2"
requests = ">=2.27.1"
pyjwt = "^2.9.0"
auth0-python = "^4.7.2"
amplitude-analytics = "^1.1.4"
certifi = "^2024.8.30"
safetensors = "^0.4.5"
openai = "1.101.0"
celery = "5.3.6"
diffusers = "^0.30.3"
accelerate = "^1.0.0"
redis = "^5.2.1"
tiktoken = "^0.11.0"
# grpcio = ">=1.53.0,<1.64.0"
pillow = "^10.4.0"
werkzeug = "3.0.1"
torch = "2.2"
torchvision = "0.17"
pydantic = ">=2.7.4,<3.0.0"
langchain-community = "^0.3.1"
langchain-huggingface = "^0.1.0"
spacy = "3.7.5"
pypdf = "^5.0.1"
scikit-learn = "1.5.2"
debugpy = "^1.8.6"
ollama = "^0.5.1"
langchain-experimental = "^0.3.2"
langchain-core = "^0.3.9"
python-magic = "^0.4.27"
scipy = "1.13.1"
numpy = "^1.24.0"
instructor = "^1.7.0"  # Flexible for rich 14.x
protobuf = ">=4.21.6"  # Allow 6.x for TensorLake 0.2.69
pycryptodome = "^3.20.0"
fastapi = "0.115.6"
uvicorn = "0.34.0"
httpx = "^0.28.1"
transformers = "^4.53.0"
sentencepiece = "0.1.99"
aiofiles = "*"
python-multipart = "*"
stripe = "11.3.0"
libmagic = "1.0"
pytest = "^8.3.4"
azure-monitor-opentelemetry = "^1.2.1"
backoff = "2.2.1"
email-validator = "^2.2.0"
pyyaml = "^6.0.2"
groq = "^0.26.0"
pymongo = "^4.13.0"
qdrant-client = "<1.8.0"
grpcio-tools = ">=1.54.0"  # Allow 1.75.x for TensorLake 0.2.69
cachetools = "^6.1.0"
fastmcp = "^2.1.0"  # or the latest version
azure-servicebus = "^7.11.0"
neo4j = "5.27.0"
en-core-web-sm = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl"}
posthog = "^3.0.0"  # Self-hostable analytics for open source
# Document processing providers
google-genai = "^1.46.0"         # NEW Gemini SDK - NO protobuf conflicts! https://googleapis.github.io/python-genai/
# google-generativeai = "^0.8.0"  # OLD (deprecated) - Conflicts with TensorLake, support ends Aug 2025
pymupdf = "^1.26.6"              # PyMuPDF - required for Gemini PDF processing
reducto = "*"                    # Reducto AI - use latest available
tensorlake = "^0.2.69"           # TensorLake.ai SDK - NOW COMPATIBLE with new google-genai!
temporalio = "^1.8.0"            # Temporal workflows for durable processing (1.8+ required for VersioningIntent)
pinecone = "^7.3.0"
motor = "^3.7.1"
reductoai = "^0.12.0"
# paddleocr = "^3.3.0"  # BLOCKED: Incompatible with tensorlake (requires PyYAML==6.0.2, tensorlake needs 6.0.3+)
python-toon = "^0.1.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-mock = "^3.14.0"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.25.0"
fakeredis = "^2.21.3"
asgi-lifespan = "^2.1.0"
pytest-timeout = "^2.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
